COM-1806: remove redondant code
