+++ src/AppContainer/index.tsx
@@ -16,8 +16,8 @@ import Authentication from '../screens/Authentication';
 import EmailForm from '../screens/EmailForm';
 import CreateAccount from '../screens/CreateAccount';
 import Catalog from '../screens/explore/Catalog';
-import BlendedAbout from '../screens/explore/About/BlendedAbout';
-import ElearningAbout from '../screens/explore/About/ELearningAbout';
+import BlendedAbout from '../screens/explore/BlendedAbout';
+import ElearningAbout from '../screens/explore/ELearningAbout';
 import CourseList from '../screens/courses/CourseList';
 import CourseProfile from '../screens/courses/CourseProfile';
 import SubProgramProfile from '../screens/courses/SubProgramProfile';
